refactor: user패키지 생성 및 관련 클래스 패키지 이동
