+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -1,6 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.user.User;
 
 public class Dealer extends User {
     private static final int DEALER_REDRAW_STANDARD = 17;
